chore(#0): Rename BeforeOffsetRelation to BeforeRelation.  Rename AfterOffsetRelation to AfterRelation
